Changed int to uint8_t in debounce_xeal.c

+++ keyboard/debounce_xeal.c
@@ -17,30 +17,30 @@ static uint8_t current_time;
 
 void debounce_matrix_init(void)
 {
-    for (int i = 0; i < MATRIX_ROWS * MATRIX_COLS; i++)
+    for (uint8_t i = 0; i < MATRIX_ROWS * MATRIX_COLS; i++)
     {
         matrix[i] = 0;
     }
-    for (int i = 0; i < MATRIX_ROWS; i++)
+    for (uint8_t i = 0; i < MATRIX_ROWS; i++)
     {
         debounced_matrix[i] = 0;
     }
 }
 
 
-inline static bool is_bouncing(int index)
+inline static bool is_bouncing(uint8_t index)
 {
     //return (matrix[index] & BOUNCE_BIT == BOUNCE_BIT);
     //return (matrix[index] & BOUNCE_BIT);
     return (matrix[index]); //since value is non-zero, this works.
 }
 
-inline static uint8_t bounce_expiry(int index)
+inline static uint8_t bounce_expiry(uint8_t index)
 {
     return (matrix[index] & DEBOUNCE_MODULO_MASK);
 }
 
-inline static void set_expiry(int index, uint8_t timeFromNow)
+inline static void set_expiry(uint8_t index, uint8_t timeFromNow)
 {
     matrix[index] = ((current_time + timeFromNow) % MAX_DEBOUNCE) ^ BOUNCE_BIT;
 }
@@ -48,7 +48,7 @@ inline static void set_expiry(int index, uint8_t timeFromNow)
 //We assume that this will be called at least once every millisecond!
 static void check_debounce_matrix(void) 
 {	    
-    for (int i = 0; i < MATRIX_ROWS * MATRIX_COLS; i++)
+    for (uint8_t i = 0; i < MATRIX_ROWS * MATRIX_COLS; i++)
     {
         if (is_bouncing(i) && current_time == bounce_expiry(i))
         {
@@ -63,7 +63,7 @@ static void check_debounce_matrix(void)
 static inline void handle_new_data(matrix_row_t* raw_values, matrix_row_t* output)
 {
 
-    for (int row = 0; row < MATRIX_ROWS; row++)
+    for (uint8_t  row = 0; row < MATRIX_ROWS; row++)
     {
         matrix_row_t raw_row = raw_values[row];
         matrix_row_t existing_row = output[row];
@@ -71,7 +71,7 @@ static inline void handle_new_data(matrix_row_t* raw_values, matrix_row_t* outpu
         uint8_t matrix_index = row * MATRIX_COLS;
 
         if (raw_row != existing_row) { //quick check for change.
-            for (int col = 0; col < MATRIX_COLS; col++)
+            for (uint8_t  col = 0; col < MATRIX_COLS; col++)
             {
                 bool existing_value = CHECK_BIT(existing_row, col);
                 bool new_value = CHECK_BIT(raw_row, col);
